Update response_structs.rs OpenOrder oid (#5)

+++ src/info/response_structs.rs
@@ -15,7 +15,7 @@ pub struct UserStateResponse {
 pub struct OpenOrdersResponse {
     pub coin: String,
     pub limit_px: String,
-    pub oid: i32,
+    pub oid: u64,
     pub side: String,
     pub sz: String,
     pub timestamp: u64,
